+++ R/tidyverse.R
@@ -83,6 +83,19 @@ select.stars_proxy = function(.data, ...) {
 	collect(.data, match.call(), "select", ".data", env = environment())
 }
 
+#' @name dplyr
+rename.stars <- function(.data, ...) {
+    if (!requireNamespace("dplyr", quietly = TRUE))
+        stop("package dplyr required, please install it first") # nocov
+    ret <- dplyr::rename(to_df(.data), ...)
+	st_as_stars(set_dim(ret, dim(.data)), dimensions = st_dimensions(.data))
+}
+
+#' @name dplyr
+rename.stars_proxy = function(.data, ...) {
+	collect(.data, match.call(), "rename", ".data", env = environment())
+}
+
 #' @param var see \link[dplyr]{pull}
 #' @name dplyr
 pull.stars = function (.data, var = -1) {
@@ -269,6 +282,8 @@ register_all_s3_methods = function() {
 	register_s3_method("dplyr", "mutate", "stars_proxy")
 	register_s3_method("dplyr", "pull", "stars")
 	register_s3_method("dplyr", "pull", "stars_proxy")
+	register_s3_method("dplyr", "rename", "stars")
+	register_s3_method("dplyr", "rename", "stars_proxy")
 	register_s3_method("dplyr", "slice", "stars")
 	register_s3_method("dplyr", "slice", "stars_proxy")
 	register_s3_method("dplyr", "transmute", "stars")